chore: remove old Stage reconciler
